2024.10
Updates

* Information about content reporting added to documentation
* Improved error handling for archiving mods with dependents, and already-redeemed email authentication codes
* ListUserCreatedModsAsync will now update the client-side cache of mod objects
* When submitting new mod files, file paths will be checked for maximum path length limits
* Modio::ErrorCode is now marked `nodiscard`. If you have enabled warnings as errors and are not consuming error codes being returned from SDK functions, you will need to either inspect those error codes or explicitly discard them with a cast to void.

+++ modio/modio/core/ModioMetricsService.h
@@ -116,7 +116,7 @@ namespace Modio
 			std::string ModIdsString;
 			std::vector<ModID> ModIds;
 			time_t SessionStartTime;
-			int CurrentSessionOrderId;
+			uint64_t CurrentSessionOrderId;
 
 			time_t LastHeartbeatPerformed;
